+++ bc/recruitment_api/management/commands/import_jobs.py
@@ -5,7 +5,7 @@ from django.utils.timezone import now
 from bc.documents.models import CustomDocument
 from bc.recruitment.models import TalentLinkJob
 from bc.recruitment_api.client import get_client
-from bc.recruitment_api.utils import update_job_from_ad
+from bc.recruitment_api.utils import delete_jobs, update_job_from_ad
 
 
 class Command(BaseCommand):
@@ -27,6 +27,7 @@ class Command(BaseCommand):
         num_created = 0
         doc_imported = 0
         errors = []
+        import_timestamp = now()
         while results:
             self.stdout.write(f"Fetching page {page}")
             response = client.service.getAdvertisementsByPage(
@@ -49,7 +50,7 @@ class Command(BaseCommand):
                         job = update_job_from_ad(
                             job,
                             ad,
-                            defaults={"last_imported": now()},
+                            defaults={"last_imported": import_timestamp},
                             import_categories=options["import_categories"],
                         )
                     except Exception as e:
@@ -97,10 +98,19 @@ class Command(BaseCommand):
                             )
                             errors.append(msg)
 
+        # Check for outdated jobs
+        num_deleted = 0
+        try:
+            num_deleted = delete_jobs(import_timestamp)
+        except Exception as e:
+            msg = f"Error occurred while deleting jobs:\n" + str(e)
+            errors.append(msg)
+
         self.stdout.write("No more results")
         self.stdout.write(f"{num_updated} existing jobs updated")
         self.stdout.write(f"{num_created} new jobs created")
         self.stdout.write(f"{doc_imported} new documents imported")
+        self.stdout.write(f"{num_deleted} jobs deleted")
         self.stdout.write(self.style.ERROR(f"{len(errors)} errors"))
         for error in errors:
             self.stdout.write(self.style.ERROR(msg))