+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - 114 StreamField Button Link block (BE)
 - 125 Job attachments
+- 148 Delete outdated jobs when running import command
 
 ## 9.0 (2020-03-31)